+++ pbx_test2/a/ContentView.swift
@@ -1,0 +1,20 @@
+//
+//  ContentView.swift
+//  pbx_test2
+//
+//  Created by 강한결 on 10/8/24.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+   var body: some View {
+      VStack {
+         Image(systemName: "globe")
+            .imageScale(.large)
+            .foregroundStyle(.tint)
+         Text("Hello, a!")
+      }
+      .padding()
+   }
+}
